third_party/blink/web_tests/platform/generic/fast/forms/form-input-named-arguments-expected.txt

PASS clicked is true
PASS successfullyParsed is true

TEST COMPLETE
